:bug: Fix old since date issue (#7671)
* :construction: Move code to an abstract base clase so it can be reused

* :construction: Add query to reset local state

* :sparkles: Handle 400 error code when pulling data
