+++ Toggl.Storage.Realm/Queries/RealmQueryFactory.cs
@@ -31,5 +31,8 @@ namespace Toggl.Storage.Realm.Queries
 
         public IQuery<Unit> MigrateBackToOldSyncing()
             => new MigrateBackToOldSyncingQuery(realmProvider);
+
+        public IQuery<Unit> ResetLocalState(Networking.Sync.Pull.IResponse response)
+            => new ResetLocalStateQuery(realmProvider, currentTimeProvider, response);
     }
 }
